fixed progress bar updating in realtime
